Add active accounts page and account deactivation feature
Created a new page to display all active accounts. Functionalities added include fetching of active accounts from the client and rendering them in a table. Also, implemented an Account Deactivation feature that allows deactivation of an account via a button click, after which the page auto-refreshes to display the updated list of active accounts.